Add function to compute the argument of a complex number
Pairs in Asymptote behave like complex numbers. There is a function to obtain
the absolute value of a complex number but there doesn't seem to be a function
to compute the argument of a complex function. This adds such a funciton. It
allows you to calculate the principal value of a complex number, either as a
value between -pi and pi or as a value between 0 and 2*pi.