Add option to insert only specific labels on the coordinate scale
Until now, you could decide either to insert no labels at all or to label all
points on the scale. In certain situations it is visually appealing to only
label some of the points on the scale.
To accomplish this, this adds an argument for each axis to the
draw_coordinates function. This argument takes an array which contains all the
labels which should be drawn for that axis.